Add espanso to config.
had to manage this from within awesome rather than nixos, which is
annoying. there's a bug with the provided systemd --user file that you
work around by calling the program from user space.

this is a brittle config, but i'll take actually having the program
work over not using it at all. the program is cool.